requirements.txt
tensorflow-cpu==2.18.0
streamlit
pillow
numpy
python-dotenv
pandas